structure of userReview components changes
