+++ src/App.tsx
@@ -4,7 +4,6 @@ import { BrowserRouter as Router, Route, Routes } from 'react-router-dom'
 import './App.css'
 const queryClient = new QueryClient()
 
-import FavoriteFood from './components/pages/FavoriteFood'
 import FavoriteFoods from './components/pages/FavoriteFoods'
 import Friends from './components/pages/Friends'
 import Home from './components/pages/Home'
@@ -13,10 +12,10 @@ import NewReviewForm from './components/pages/NewReviewForm'
 import Profile from './components/pages/Profile'
 import Register from './components/pages/Register'
 import Restaurant from './components/pages/Restaurant'
-import UserReviews from './components/pages/UserReviews'
 import Navbar from './components/partials/NavBar'
 import PrivateRoute from './components/partials/PrivateRoute'
 import { UserProvider } from './components/partials/UserContext'
+import UserReviews from './components/partials/UserReviewsByFood'
 
 function App() {
   const theme = createTheme()
@@ -64,14 +63,14 @@ function App() {
                     </PrivateRoute>
                   }
                 />
-                <Route
-                  path='/:username/favorites/:favorite'
+                {/* <Route
+                  path='/:username/reviews/:favorite'
                   element={
                     <PrivateRoute>
-                      <FavoriteFood />
+                      <UserReviews />
                     </PrivateRoute>
                   }
-                />
+                /> */}
                 <Route path='/restaurants/:restaurantId' element={<Restaurant />} />
                 <Route path='/register' element={<Register />} />
                 <Route path='/login' element={<Login />} />
